Organization changes
Renamed a lot of files into a simplier organization.  Fixed the tests to
use the new structure and removed a ton of unused code from the
fast-html-parser port.  Was able to remove the browserify bundle wrapper
from it as well.
