+++ .jshintrc
@@ -1,4 +1,5 @@
 {
+  "loopfunc": true,
   "esnext": true,
   "evil": true,
   "shadow": true,
