--- lib/diff/make_node.js
@@ -1,84 +1,0 @@
-import { pools } from '../util/pools';
-
-var push = Array.prototype.push;
-var nodes = makeNode.nodes = {};
-
-/**
- * Converts a live node into a virtual node.
- *
- * @param node
- * @return
- */
-function makeNode(node) {
-  // If this node has already been converted, do not attempt to convert again.
-  if (node && node.__node__) {
-    return node.__node__;
-  }
-
-  var nodeType = node.nodeType;
-  var nodeValue = node.nodeValue;
-
-  if (!nodeType || nodeType === 2 || nodeType === 4 || nodeType === 8) {
-    return false;
-  }
-
-  if (nodeType === 3 && !nodeValue.trim()) {
-    return false;
-  }
-
-  // Virtual representation of a node, containing only the data we wish to
-  // diff and patch.
-  var entry = {};
-
-  // Cache the element in the ids.
-  var id = pools.uuid.get();
-
-  // Add to internal lookup.
-  nodes[id] = node;
-
-  // Save a reference to this object.
-  node.__node__ = entry;
-
-  entry.element = id;
-  entry.nodeName = node.nodeName.toLowerCase();
-  entry.nodeValue = nodeValue;
-  entry.childNodes = [];
-  entry.attributes = [];
-
-  // Collect attributes.
-  var attributes = node.attributes;
-
-  // If the element has no attributes, skip over.
-  if (attributes) {
-    var attributesLength = attributes.length;
-
-    if (attributesLength) {
-      for (var i = 0; i < attributesLength; i++) {
-        push.call(entry.attributes, {
-          name: attributes[i].name,
-          value: attributes[i].value
-        });
-      }
-    }
-  }
-
-  // Collect childNodes.
-  var childNodes = node.childNodes;
-  var childNodesLength = node.childNodes.length;
-  var newNode = null;
-
-  // If the element has child nodes, convert them all to virtual nodes.
-  if (node.nodeType !== 3 && childNodes) {
-    for (var i = 0; i < childNodesLength; i++) {
-      newNode = makeNode(childNodes[i]);
-
-      if (newNode) {
-        entry.childNodes.push(newNode);
-      }
-    }
-  }
-
-  return entry;
-}
-
-export default makeNode;
