+++ lib/element.js
@@ -1,4 +1,4 @@
-import makeNode from './make_node';
+import { makeNode } from './node';
 import * as svg from './svg';
 
 /**
@@ -7,7 +7,7 @@ import * as svg from './svg';
  * @param descriptor
  * @return
  */
-function makeElement(descriptor) {
+export function makeElement(descriptor) {
   var element = null;
   var isSvg = false;
 
@@ -47,5 +47,3 @@ function makeElement(descriptor) {
 
   return element;
 }
-
-export default makeElement;
