+++ lib/index.js
@@ -1,4 +1,4 @@
-import patch from './diff/patch_node';
+import { patchNode } from './node';
 
 /**
  * outer
@@ -10,7 +10,7 @@ import patch from './diff/patch_node';
  */
 export function outerHTML(element, markup='', options={}) {
   options.inner = false;
-  patch(element, markup, options);
+  patchNode(element, markup, options);
 }
 
 /**
@@ -23,7 +23,7 @@ export function outerHTML(element, markup='', options={}) {
  */
 export function innerHTML(element, markup='', options={}) {
   options.inner = true;
-  patch(element, markup, options);
+  patchNode(element, markup, options);
 }
 
 /**
@@ -36,7 +36,7 @@ export function innerHTML(element, markup='', options={}) {
  */
 export function element(element, newElement, options={}) {
   options.inner = false;
-  patch(element, newElement, options);
+  patchNode(element, newElement, options);
 }
 
 /**
@@ -48,7 +48,7 @@ export function enableProllyfill() {
   Object.defineProperty(Element.prototype, 'addTransitionState', {
     configurable: true,
 
-    value: function(name, callback) {
+    value(name, callback) {
       var states = this._transitionStates = this._transitionStates || {};
 
       states[name] = states[name] || [];
@@ -60,7 +60,7 @@ export function enableProllyfill() {
   Object.defineProperty(Element.prototype, 'removeTransitionState', {
     configurable: true,
 
-    value: function(name, callback) {
+    value(name, callback) {
       var states = this._transitionStates = this._transitionStates || {};
 
       states[name] = states[name] || [];
@@ -77,7 +77,7 @@ export function enableProllyfill() {
   Object.defineProperty(Element.prototype, 'diffInnerHTML', {
     configurable: true,
 
-    set: function(newHTML) {
+    set(newHTML) {
       innerHTML(this, newHTML);
     }
   });
@@ -85,7 +85,7 @@ export function enableProllyfill() {
   Object.defineProperty(Element.prototype, 'diffOuterHTML', {
     configurable: true,
 
-    set: function(newHTML) {
+    set(newHTML) {
       outerHTML(this, newHTML);
     }
   });
@@ -93,7 +93,7 @@ export function enableProllyfill() {
   Object.defineProperty(Element.prototype, 'diffElement', {
     configurable: true,
 
-    value: function(newElement) {
+    value(newElement) {
       element(this, newElement);
     }
   });
