+++ lib/svg.js
