+++ package.json
@@ -18,16 +18,17 @@
   },
   "devDependencies": {
     "babelify": "^6.1.3",
-    "browserify": "^9.0.8",
+    "browserify": "^11.0.1",
     "browserify-istanbul": "^0.2.1",
     "derequire": "^2.0.2",
     "http-server": "^0.8.0",
-    "istanbul": "^0.3.13",
-    "jshint": "^2.7.0",
-    "mocha": "^2.2.4",
-    "mocha-phantomjs": "^3.5.2",
-    "mocha-phantomjs-istanbul": "0.0.2",
+    "istanbul": "^0.3.18",
+    "jshint": "^2.8.0",
+    "mocha": "^2.2.5",
+    "mocha-phantomjs": "^3.6.0",
+    "mocha-phantomjs-istanbul": "^0.0.2",
     "promise-polyfill": "^2.1.0",
-    "watchify": "^3.1.2"
+    "watchify": "^3.3.1",
+    "weakmap": "^0.0.6"
   }
 }
