+++ test/playground.html
@@ -8,16 +8,5 @@
 </head>
 <body>
   <div></div>
-
-  <script src="../dist/diffhtml.js"></script>
-
-  <script>
-    // Test attribute changes.
-    var div = document.querySelector('div');
-
-    //setInterval(function() {
-      div.outerDiffHTML = '<div class="' + Date.now() + '"><p>' + Date.now() + '</p></div>';
-    //}, 0);
-  </script>
 </body>
 </html>
